New designed navbar added
