+++ src/app/page.tsx
@@ -1,3 +1,4 @@
+"use client";
 import {
   ARSIndustry,
   Awards,
