+++ src/sections/home/solutions.tsx
@@ -1,3 +1,4 @@
+"use client";
 import { solutionsData } from "@/data";
 import React from "react";
