+++ src/sections/index.ts
@@ -1,6 +1,5 @@
 import Footer from "./footer/footer";
 import Header from "./header/header";
-
 // Home sections
 import Hero from "./home/hero";
 import OurMission from "./home/mission";
